requirements.txt
Flask==1.1.2
Flask-Migrate==2.6.0
psycopg2==2.8.6
Flask-Testing==0.8.1
Flask-JWT-Extended==3.25.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
requests==2.25.1